harness: a phase on a Time was watching the high word, which is always zero
Found while preparing W2's deploy, before it reached the rig. It would have made
the whole feature silently do nothing.

A Time occupies TWO registers, high-word-first, and ResultRegisterOf returns a
signal's FIRST register - so a phase condition on a window clock was watching the
HIGH word. Any elapsed value under 65.536 s leaves that at 0 permanently while
every millisecond lands in the second register. A `Below` trigger on such a signal
is therefore satisfied on the FIRST POLL of every wave, reporting a freshly
restarted window including on waves where no window is running at all.

This is not a hypothesis: the same high-word blindness is already recorded in this
system against a mirrored timer, where a declaration written against the high word
alone could not see the signal move and presented as an unrelated failure two
registers along.

PhaseCondition now carries the signal's WIDTH and word order, assembles the value
across both registers before comparing, and checks that the WHOLE signal fits the
band rather than just its first register - a Time straddling the end would have
read its high word and indexed past the array for its low one. The width comes
from the binding's own declaration, never from an assumed 1, and a width the map
cannot produce is refused rather than truncated.

The threshold widens to uint with it: a two-minute window is 120 000 ms, which does
not fit the ushort the API previously took.

The test carries its own control - the same registers read at width 1 return 0 and
the trigger fires, which is the bug, beside width 2 returning 30 000 and correctly
not firing. And it asserts a genuinely fresh window IS still caught, because the
other way a check like this goes quiet is by disabling the trigger outright.

2,233 passing, 0 failing, 0 warnings.

+++ src/harness/Harness.Gate/BindingDocument.cs
@@ -219,7 +219,7 @@ public sealed class SlotBindingDocument
     /// The comparison value for <c>Below</c> and <c>AtOrAbove</c>; refused with <c>Decreases</c>, which
     /// compares against the previous sample rather than a number.
     /// </summary>
-    public ushort? PhaseThreshold { get; set; }
+    public uint? PhaseThreshold { get; set; }
 
     /// <summary>
     /// A signal that must be SET for a phase sample to count. <b>Required in practice for any clock that
